(refactor and ui update mode) : implement filter logic
